Refactor layout structure in session page; adjust column spans for LeftSection, MiddleSection, and RightSection components
MADE UI CHANGES
