+++ twai/components/SessionPageComponents/left-section.js
@@ -3,26 +3,28 @@
 import { useAppContext } from "../../context/AppContext"
 import { useState, useRef, useEffect } from "react"
 import { Button } from "@/components/ui/button"
-import { MessageSquare, ScrollText, Send, Trash, Keyboard, Mic } from "lucide-react"
+import { MessageSquare, ScrollText, Send, Trash, Keyboard, LogOut } from "lucide-react"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { ScrollArea } from "@/components/ui/scroll-area"
-import { useParams } from "next/navigation"
+import { useParams, useRouter } from "next/navigation"
 import { appendConversation } from "../../app/session/[sessionId]/actions"
 import { Textarea } from "@/components/ui/textarea"
 import { Switch } from "@/components/ui/switch"
 import { Label } from "@/components/ui/label"
 import { cn } from "@/lib/utils"
-import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs"
+import MicrophoneButton from "@/components/SessionPageComponents/microphoneButton"
+import CaptureScreenButton from "@/components/SessionPageComponents/captureScreenButton"
 
 export default function LeftSection() {
-  const { wholeConversation, setWholeConversation, micPartialTranscript, capturePartialTranscript } = useAppContext()
+  const { wholeConversation, setWholeConversation, videoRef, stream } = useAppContext()
   const { sessionId } = useParams()
+  const router = useRouter()
   const [autoScroll, setAutoScroll] = useState(true)
   const [manualInput, setManualInput] = useState("")
   const [isUser, setIsUser] = useState(false)
+  const [showInput, setShowInput] = useState(false)
   const scrollAreaRef = useRef(null)
   const conversationEndRef = useRef(null)
-  const [activeTab, setActiveTab] = useState("input")
 
   const handleClearConversation = async () => {
     const tempConversation = [...wholeConversation]
@@ -35,7 +37,13 @@ export default function LeftSection() {
       const newMessage = isUser ? { user: manualInput } : { other: manualInput }
       setWholeConversation([...wholeConversation, newMessage])
       setManualInput("")
+      setShowInput(false) // Hide input after sending
     }
+  }
+
+  const handleExit = async () => {
+    await appendConversation({ sessionId: sessionId, newMessages: wholeConversation })
+    router.push("/dashboard")
   }
 
   // Auto-scroll effect
@@ -45,25 +53,32 @@ export default function LeftSection() {
     }
   }, [wholeConversation, autoScroll])
 
-  useEffect(() => {
-    // Auto-scroll transcript containers when content changes
-    const captureContainer = document.getElementById("capture-transcript-container")
-    const micContainer = document.getElementById("mic-transcript-container")
-
-    // Use requestAnimationFrame to ensure the DOM has updated before scrolling
-    requestAnimationFrame(() => {
-      if (captureContainer) {
-        captureContainer.scrollTop = captureContainer.scrollHeight
-      }
-
-      if (micContainer) {
-        micContainer.scrollTop = micContainer.scrollHeight
-      }
-    })
-  }, [capturePartialTranscript, micPartialTranscript])
-
   return (
-    <div className="h-[calc(100vh-120px)] flex flex-col">
+    <div className="h-[calc(100vh-32px)] flex flex-col gap-4">
+      {/* Screen Capture Card */}
+      <Card className="border shadow-sm">
+        <CardHeader className="p-4 pb-2 flex flex-row items-center justify-between">
+          <CardTitle className="text-lg font-medium">Screen Capture</CardTitle>
+          <Button onClick={handleExit} variant="destructive" size="sm" className="flex items-center gap-1">
+            <LogOut className="h-4 w-4" />
+            Exit Session
+          </Button>
+        </CardHeader>
+        <CardContent className="p-4 pt-0">
+          {/* Video Section */}
+          <div className="aspect-video bg-gray-800 rounded-lg overflow-hidden mb-3">
+            <video ref={videoRef} autoPlay playsInline className="w-full h-full object-cover" />
+          </div>
+
+          {/* Buttons Section */}
+          <div className="flex flex-col sm:flex-row gap-2">
+            <MicrophoneButton />
+            <CaptureScreenButton />
+          </div>
+        </CardContent>
+      </Card>
+
+      {/* Conversation Card */}
       <Card className="border shadow-sm flex-1 flex flex-col overflow-hidden">
         <CardHeader className="p-4 pb-2 flex flex-row items-center justify-between shrink-0">
           <div className="flex items-center gap-2">
@@ -122,71 +137,55 @@ export default function LeftSection() {
             </ScrollArea>
           </div>
 
-          {/* Togglable Transcription/Input Section */}
-          <div className="border-t pt-3 mt-3">
-            <Tabs value={activeTab} onValueChange={setActiveTab} className="w-full">
-              <TabsList className="grid grid-cols-2 mb-2">
-                <TabsTrigger value="transcripts" className="flex items-center gap-1">
-                  <Mic className="h-3.5 w-3.5" />
-                  <span>Transcripts</span>
-                </TabsTrigger>
-                <TabsTrigger value="input" className="flex items-center gap-1">
-                  <Keyboard className="h-3.5 w-3.5" />
-                  <span>Manual Input</span>
-                </TabsTrigger>
-              </TabsList>
+          {/* Add Message Button */}
+          {!showInput && (
+            <Button variant="outline" className="mt-3" onClick={() => setShowInput(true)}>
+              <Keyboard className="h-4 w-4 mr-2" />
+              Add Message
+            </Button>
+          )}
 
-              <TabsContent value="transcripts" className="mt-0">
-                <div className="grid grid-cols-2 gap-3">
-                  <div className="bg-gray-100 p-2 rounded-lg h-16 overflow-hidden relative">
-                    <h3 className="text-xs font-medium text-muted-foreground mb-1">Captured Audio</h3>
-                    <div className="text-sm text-gray-800 overflow-auto h-8" id="capture-transcript-container">
-                      {capturePartialTranscript || "No transcription available..."}
-                    </div>
-                  </div>
-                  <div className="bg-gray-100 p-2 rounded-lg h-16 overflow-hidden relative">
-                    <h3 className="text-xs font-medium text-muted-foreground mb-1">Mic Audio</h3>
-                    <div className="text-sm text-gray-800 overflow-auto h-8" id="mic-transcript-container">
-                      {micPartialTranscript || "No transcription available..."}
-                    </div>
-                  </div>
-                </div>
-              </TabsContent>
-
-              <TabsContent value="input" className="mt-0">
-                <div className="flex flex-col space-y-2">
-                  <div className="flex items-center justify-between">
-                    <Label htmlFor="message-type" className="text-sm">
-                      Message as: {isUser ? "You" : "Other"}
-                    </Label>
+          {/* Manual Input Section - Collapsible */}
+          {showInput && (
+            <div className="border-t pt-3 mt-3 animate-in slide-in-from-bottom duration-200">
+              <div className="flex flex-col space-y-2">
+                <div className="flex items-center justify-between">
+                  <Label htmlFor="message-type" className="text-sm">
+                    Message as: {isUser ? "You" : "Other"}
+                  </Label>
+                  <div className="flex items-center gap-2">
                     <Switch id="message-type" checked={isUser} onCheckedChange={setIsUser} />
-                  </div>
-                  <div className="flex space-x-2">
-                    <Textarea
-                      placeholder="Type message to add..."
-                      value={manualInput}
-                      onChange={(e) => setManualInput(e.target.value)}
-                      className="min-h-[60px] flex-1"
-                      onKeyDown={(e) => {
-                        if (e.key === "Enter" && !e.shiftKey) {
-                          e.preventDefault()
-                          handleAddConversation()
-                        }
-                      }}
-                    />
-                    <Button
-                      onClick={handleAddConversation}
-                      disabled={!manualInput.trim()}
-                      size="icon"
-                      className="self-end"
-                    >
-                      <Send className="h-4 w-4" />
+                    <Button variant="ghost" size="sm" onClick={() => setShowInput(false)} className="h-8 w-8 p-0">
+                      <Trash className="h-4 w-4" />
                     </Button>
                   </div>
                 </div>
-              </TabsContent>
-            </Tabs>
-          </div>
+                <div className="flex space-x-2">
+                  <Textarea
+                    placeholder="Type message to add..."
+                    value={manualInput}
+                    onChange={(e) => setManualInput(e.target.value)}
+                    className="min-h-[60px] flex-1"
+                    onKeyDown={(e) => {
+                      if (e.key === "Enter" && !e.shiftKey) {
+                        e.preventDefault()
+                        handleAddConversation()
+                      }
+                    }}
+                    autoFocus
+                  />
+                  <Button
+                    onClick={handleAddConversation}
+                    disabled={!manualInput.trim()}
+                    size="icon"
+                    className="self-end"
+                  >
+                    <Send className="h-4 w-4" />
+                  </Button>
+                </div>
+              </div>
+            </div>
+          )}
         </CardContent>
       </Card>
     </div>
