+++ twai/components/SessionPageComponents/right-section.js
@@ -1,20 +1,18 @@
 "use client"
 
-import { Button } from "@/components/ui/button"
-import { Bot, X, Send, CloudUpload, Trash, Search, BarChart2, Highlighter, ScrollText } from "lucide-react"
 import { useAppContext } from "../../context/AppContext"
-import { get_AI_Help } from "../../lib/sessionActions"
+import { useState } from "react"
+import { Button } from "@/components/ui/button"
+import { BookOpen, FileText, Clock, BarChart2, X, ExternalLink } from "lucide-react"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
 import { ScrollArea } from "@/components/ui/scroll-area"
-import { appendConversation } from "../../app/session/[sessionId]/actions"
 import { useParams } from "next/navigation"
-import { useState, useRef, useEffect } from "react"
-import { Textarea } from "@/components/ui/textarea"
-import { Checkbox } from "@/components/ui/checkbox"
-import ReactMarkdown from "react-markdown"
-import { unstable_noStore as noStore } from 'next/cache';
+import { appendConversation } from "../../app/session/[sessionId]/actions"
+import { unstable_noStore as noStore } from "next/cache"
+import { TooltipProvider, Tooltip, TooltipTrigger, TooltipContent } from "@/components/ui/tooltip"
+
 export default function RightSection() {
-  noStore();
+  noStore()
   const {
     isProcessing,
     setIsProcessing,
@@ -25,344 +23,315 @@ export default function RightSection() {
     chatMessages,
     wholeConversation,
     enableWebSearch,
-    setEnableWebSearch,
-    showGraph,
-    setShowGraph,
     usedCitations,
     setUsedCitations,
     useHighlightedText,
-    setUseHighlightedText,
     copiedText,
-    userInput,
-    setUserInput,
-    setGraphImage
+    setCopiedText,
+    graphImage,
+    setGraphImage,
+    setShowGraph,
+    showGraph,
   } = useAppContext()
 
   const { sessionId } = useParams()
-  const [autoScroll, setAutoScroll] = useState(true)
-  const chatEndRef = useRef(null)
-  const [image, setImage] = useState(null)
+  const [isGraphVisible, setIsGraphVisible] = useState(false)
 
+  const toggleGraphVisibility = () => {
+    setIsGraphVisible(!isGraphVisible)
+  }
 
-
-  const handleSendMessage = async () => {
+  const handleAIAnswer = async (requestType) => {
+    if (isProcessing) return
     setGraphImage(null)
-    if (userInput.trim()) {
-      setIsProcessing(true)
-      setChatMessages((prev) => [...prev, { text: userInput, sender: "user" }])
-      setChatMessages((prev) => [...prev, { text: "Thinking...", sender: "ai"} ]);
-      setUserInput("")
+    setIsProcessing(true)
+    setChatMessages([...chatMessages, { text: "Thinking...", sender: "ai" }])
 
-      const formData = new FormData()
-      formData.append("user_input", userInput)
-      formData.append("use_web", enableWebSearch)
-      formData.append("use_graph", showGraph)
-      formData.append("sessionId", sessionId)
-      if (image) {
-        formData.append("uploaded_file", image)
-      }
+    try {
+      // Step 1: Extract the query using OpenAI on the frontend
+      const tempconv = [...wholeConversation]
+      setWholeConversation([])
 
-      if (wholeConversation) {
-        formData.append("raw_Conversation", JSON.stringify(wholeConversation))
-      }
-
-      // Send the request to the server-side route
-      const response = await fetch("/api/chat_Jamie_AI", {
+      // Step 3: Send the extracted query to the backend for processing
+      const response = await fetch("/api/get_AI_Help", {
         method: "POST",
-        body: formData,
+        headers: { "Content-Type": "application/json" },
+        body: JSON.stringify({
+          conversation: tempconv,
+          use_web: enableWebSearch,
+          requestType,
+          useHighlightedText,
+          copiedText,
+          sessionId,
+        }),
       })
 
-      if (!response) throw new Error("AI response is undefined");
-  
-      const reader = response.body.getReader();
-      const decoder = new TextDecoder();
-      let buffer = '';
+      if (!response.ok) throw new Error(`Server responded with ${response.status}`)
+
+      const reader = response.body?.getReader()
+      if (!reader) throw new Error("Streaming not supported")
+
+      const decoder = new TextDecoder()
+      let buffer = ""
+
       while (true) {
-        const { done, value } = await reader.read();
-        if (done) break;
-        buffer += decoder.decode(value, { stream: true });
+        const { done, value } = await reader.read()
+        if (done) break
 
-        // Ensure we process full JSON objects by splitting on newline (\n)
-        const lines = buffer.split("\n").filter(line => line.trim() !== "");
-        // buffer = lines.pop() || ""; // Save the last (possibly incomplete) line for next read
+        buffer += decoder.decode(value, { stream: true })
 
+        const lines = buffer.split("\n").filter((line) => line.trim() !== "")
 
-
-        console.log("Lines:", lines);
-        console.log("Remaining Buffer:", buffer);
-        console.log("Buffer split:",buffer.split("\n").filter(line => line.trim() !== ""))
         for (const line of lines) {
           try {
-              const h = JSON.parse(line);
-              buffer = lines.slice(1).join() || "";
-              // if (h.query) {
-              //   setChatMessages((prev) => [
-              //     ...prev.filter((msg) => msg.text !== "Thinking..."),
-              //     { text: h.query, sender: "user" },
-              //   ]);
-              // }
-              if (h.result) {
-                setChatMessages((prev) => [
-                  ...prev.filter((msg) => msg.text !== "Thinking..."),
-                  { text: h.result, sender: "ai" },
-                ]);
-              }
-              if (h.used_citations) {
-                setUsedCitations(
-                  Object.entries(h.used_citations).map(([key, value]) => ({
-                    id: key,
-                    ...value,
-                  }))
-                );
-              }
-              if(h.graph){
-                setGraphImage(h.graph)
-                setShowGraph(true)
-              }
+            const h = JSON.parse(line)
+            buffer = lines.slice(1).join() || ""
+
+            if (h.query) {
+              setChatMessages((prev) => [
+                ...prev.filter((msg) => msg.text !== "Thinking..."),
+                { text: h.query, sender: "user" },
+                { text: "Thinking...", sender: "ai" },
+              ])
+            }
+
+            if (h.result) {
+              setChatMessages((prev) => [
+                ...prev.filter((msg) => msg.text !== "Thinking..."),
+                { text: h.result, sender: "ai" },
+              ])
+            }
+
+            if (h.used_citations) {
+              setUsedCitations(
+                Object.entries(h.used_citations).map(([key, value]) => ({
+                  id: key,
+                  ...value,
+                })),
+              )
+            }
+
+            if (h.graph) {
+              setGraphImage(h.graph)
+              setShowGraph(true)
+            }
           } catch (error) {
-              console.error('JSON Parsing Error:', error);
+            console.warn("Streaming JSON parse error:", error)
           }
+        }
       }
 
+      setCopiedText("")
+      setChatMessages((prev) => [...prev.filter((msg) => msg.text !== "Thinking...")])
 
-      
-    }
-    setChatMessages((prev) => [
-      ...prev.filter((msg) => msg.text !== "Thinking...")
-    ]);
+      const appending = await appendConversation({ sessionId, newMessages: tempconv })
 
-      // Update chat messages with AI response
-      // setChatMessages((prev) => [
-      //   ...prev.filter((msg) => msg.text !== "Thinking..."),
-      //   { text: data.answer, sender: "ai" },
-      // ])
-
-      // if (data.used_citations) {
-      //   setUsedCitations(
-      //     Object.entries(data.used_citations).map(([key, value]) => ({
-      //       id: key,
-      //       ...value,
-      //     })),
-      //   )
-      // } else {
-      //   setUsedCitations([])
-      // }
+      if (appending.success) {
+        setCapturePartialTranscript("")
+        setMicPartialTranscript("")
+      } else {
+        console.log("Error appending conversation")
+        setWholeConversation(tempconv)
+      }
+    } catch (error) {
+      console.error("AI Request failed:", error)
+      setChatMessages((prev) => [
+        ...prev.filter((msg) => msg.text !== "Thinking..."),
+        { text: "An error occurred while processing your request.", sender: "ai" },
+      ])
+    } finally {
       setIsProcessing(false)
     }
   }
 
-  const handleClear = () => {
-    setChatMessages([])
-    setUserInput("")
-  }
-
-  const handleImageUpload = (e) => {
-    const file = e.target.files[0]
-    if (file && ["image/png", "image/jpeg", "image/jpg"].includes(file.type)) {
-      setImage(file)
-    } else {
-      alert("Please upload a valid image file (PNG, JPG, JPEG).")
-    }
-  }
-
-  const triggerFileInput = () => {
-    document.getElementById("imageUpload").click()
-  }
-
-  const handleRemoveImage = () => {
-    setImage(null)
-  }
-
-  // Auto-scroll effect for chat
-  useEffect(() => {
-    if (autoScroll && chatEndRef.current) {
-      chatEndRef.current.scrollIntoView({ behavior: "smooth" })
-    }
-  }, [chatMessages, autoScroll])
-
   return (
-    <div className="h-[calc(100vh-120px)] flex flex-col overflow-hidden">
-      {/* AI Chat Card - Full height */}
-      <Card className="border shadow-sm flex-1 flex flex-col overflow-hidden">
-        <CardHeader className="p-4 pb-2 flex flex-row items-center justify-between shrink-0">
-          <div className="flex items-center gap-2">
-            <Bot className="h-5 w-5 text-primary" />
-            <CardTitle className="text-lg font-medium">AI Meeting Helper</CardTitle>
-            {isProcessing && (
-    <div className="flex space-x-2">
-      <div className="w-2.5 h-2.5 bg-primary rounded-full animate-bounce"></div>
-      <div className="w-2.5 h-2.5 bg-primary rounded-full animate-bounce delay-150"></div>
-      <div className="w-2.5 h-2.5 bg-primary rounded-full animate-bounce delay-300"></div>
-    </div>
-  )}
-          </div>
-          <div className="flex items-center gap-2">
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={() => setAutoScroll(!autoScroll)}
-              className={`text-muted-foreground h-8 ${autoScroll ? "bg-primary/10 text-primary" : ""}`}
-            >
-              <ScrollText className="h-4 w-4 mr-1" />
-              {autoScroll ? "Auto-scroll On" : "Auto-scroll Off"}
-            </Button>
-            <Button
-              variant="ghost"
-              size="sm"
-              onClick={handleClear}
-              className="text-muted-foreground h-8"
-              disabled={chatMessages.length === 0}
-            >
-              <X className="h-4 w-4 mr-1" />
-              Clear Chat
-            </Button>
-          </div>
+    <div className="h-full flex flex-col gap-4">
+      {/* AI Tools Card */}
+      <Card className="border shadow-sm">
+        <CardHeader className="p-3 pb-2">
+          <CardTitle className="text-lg font-medium">AI Tools</CardTitle>
         </CardHeader>
+        <CardContent className="p-3 pt-0">
+          <div className="grid grid-cols-1 gap-2">
+            <TooltipProvider>
+              <Tooltip>
+                <TooltipTrigger asChild>
+                  <Button
+                    disabled={isProcessing}
+                    onClick={() => handleAIAnswer("help")}
+                    variant="outline"
+                    className="justify-start"
+                    size="sm"
+                  >
+                    <BookOpen className="w-4 h-4 mr-2" />
+                    AI Answer
+                  </Button>
+                </TooltipTrigger>
+                <TooltipContent>
+                  <p>Get AI assistance with your meeting questions</p>
+                </TooltipContent>
+              </Tooltip>
+            </TooltipProvider>
 
-        <CardContent className="p-4 pt-0 flex-1 flex flex-col overflow-hidden">
-          {/* Scrollable Messages Area */}
-          <div className="flex-1 overflow-hidden">
-            <ScrollArea className="h-full pr-3">
-              {chatMessages.length === 0 ? (
-                <div className="h-full flex items-center justify-center text-center p-8">
-                  <div className="max-w-sm">
-                    <Bot className="h-12 w-12 text-muted-foreground mx-auto mb-4" />
-                    <h3 className="text-lg font-medium mb-2">Welcome to AI Meeting Helper</h3>
-                    <p className="text-muted-foreground text-sm">
-                      Ask questions or upload an image to get assistance with your meeting preparation.
-                    </p>
-                  </div>
-                </div>
-              ) : (
-                <div className="space-y-4 py-2">
-                  {chatMessages.map((message, index) => (
-                    <div key={index} className={`flex ${message.sender === "user" ? "justify-end" : "justify-start"}`}>
-                      <div
-                        className={`p-3 rounded-lg max-w-[85%] ${
-                          message.sender === "user" ? "bg-primary text-primary-foreground" : "bg-muted"
-                        }`}
-                      >
-                        {message.text === "Thinking..." ? (
-                          <div className="flex items-center gap-2">
-                            <div className="animate-pulse">Thinking</div>
-                            <span className="animate-bounce">.</span>
-                            <span className="animate-bounce delay-100">.</span>
-                            <span className="animate-bounce delay-200">.</span>
-                          </div>
-                        ) : (
-                          <div className="text-sm">
-                            <ReactMarkdown>{message.text}</ReactMarkdown>
-                          </div>
-                        )}
-                      </div>
-                    </div>
-                  ))}
-                  <div ref={chatEndRef} />
-                </div>
-              )}
-            </ScrollArea>
-          </div>
+            <TooltipProvider>
+              <Tooltip>
+                <TooltipTrigger asChild>
+                  <Button
+                    disabled={isProcessing}
+                    onClick={() => handleAIAnswer("factcheck")}
+                    variant="outline"
+                    className="justify-start"
+                    size="sm"
+                  >
+                    <FileText className="w-4 h-4 mr-2" />
+                    Fact Checking
+                  </Button>
+                </TooltipTrigger>
+                <TooltipContent>
+                  <p>Verify facts in the conversation</p>
+                </TooltipContent>
+              </Tooltip>
+            </TooltipProvider>
 
-          {/* Fixed Input Area */}
-          <div className="mt-4 shrink-0">
-            {image && (
-              <div className="mb-3 p-2 bg-muted rounded-lg flex items-center">
-                <img
-                  src={URL.createObjectURL(image) || "/placeholder.svg"}
-                  alt="Uploaded Preview"
-                  className="h-12 w-12 object-cover rounded-md"
-                />
-                <span className="text-sm ml-2 flex-1 truncate">{image.name}</span>
-                <Button
-                  variant="ghost"
-                  size="sm"
-                  onClick={handleRemoveImage}
-                  className="text-destructive hover:text-destructive/90 h-8 w-8 p-0"
-                >
-                  <Trash className="h-4 w-4" />
-                </Button>
-              </div>
+            <TooltipProvider>
+              <Tooltip>
+                <TooltipTrigger asChild>
+                  <Button
+                    disabled={isProcessing}
+                    onClick={() => handleAIAnswer("summary")}
+                    variant="outline"
+                    className="justify-start"
+                    size="sm"
+                  >
+                    <Clock className="w-4 h-4 mr-2" />
+                    Summarize
+                  </Button>
+                </TooltipTrigger>
+                <TooltipContent>
+                  <p>Get a summary of the conversation so far</p>
+                </TooltipContent>
+              </Tooltip>
+            </TooltipProvider>
+
+            {/* Graph Icon - Positioned Below AI Tools */}
+            {graphImage && showGraph && (
+              <Button
+                variant="outline"
+                size="sm"
+                className="w-full flex items-center justify-start"
+                onClick={toggleGraphVisibility}
+              >
+                <BarChart2 className="h-4 w-4 mr-2" />
+                {isGraphVisible ? "Hide Graph" : "Show Graph"}
+              </Button>
             )}
-
-            <div className="flex flex-col gap-2">
-              <div className="flex gap-2">
-                <Textarea
-                  value={userInput}
-                  onChange={(e) => setUserInput(e.target.value)}
-                  placeholder="Type a message..."
-                  className="resize-none min-h-[80px]"
-                  onKeyDown={(e) => {
-                    if (e.key === "Enter" && !e.shiftKey) {
-                      e.preventDefault()
-                      handleSendMessage()
-                    }
-                  }}
-                />
-                <div className="flex flex-col gap-2">
-                  <Button disabled={isProcessing || !userInput.trim()} onClick={handleSendMessage} className="flex-1">
-                    <Send className="h-4 w-4" />
-                  </Button>
-                  <Button variant="outline" size="icon" className="flex-1" onClick={triggerFileInput}>
-                    <CloudUpload className="h-4 w-4" />
-                  </Button>
-                  <input
-                    type="file"
-                    id="imageUpload"
-                    accept="image/png, image/jpeg, image/jpg"
-                    onChange={handleImageUpload}
-                    style={{ display: "none" }}
-                  />
-                </div>
-              </div>
-
-              <div className="flex items-center justify-between text-xs text-muted-foreground mt-2">
-                <div className="flex space-x-4">
-                  <label className="flex items-center gap-2 cursor-pointer">
-                    <Checkbox
-                      checked={enableWebSearch}
-                      onCheckedChange={() => setEnableWebSearch(!enableWebSearch)}
-                      id="web-search"
-                      className={enableWebSearch ? "bg-primary text-primary-foreground" : ""}
-                    />
-                    <div className="flex items-center gap-1">
-                      <Search className={`h-4 w-4 ${enableWebSearch ? "text-primary" : "text-muted-foreground"}`} />
-                      <span className={enableWebSearch ? "text-primary font-medium" : ""}>Web Search</span>
-                    </div>
-                  </label>
-
-                  <label className="flex items-center gap-2 cursor-pointer">
-                    <Checkbox
-                      checked={showGraph}
-                      onCheckedChange={() => setShowGraph(!showGraph)}
-                      id="show-graph"
-                      className={showGraph ? "bg-primary text-primary-foreground" : ""}
-                    />
-                    <div className="flex items-center gap-1">
-                      <BarChart2 className={`h-4 w-4 ${showGraph ? "text-primary" : "text-muted-foreground"}`} />
-                      <span className={showGraph ? "text-primary font-medium" : ""}>Graph Visualization</span>
-                    </div>
-                  </label>
-                  <label className="flex items-center gap-2 cursor-pointer">
-                    <Checkbox
-                      checked={useHighlightedText}
-                      onCheckedChange={() => setUseHighlightedText(!useHighlightedText)}
-                      id="use-highlighted-text-toggle"
-                      className={useHighlightedText ? "bg-primary text-primary-foreground" : ""}
-                    />
-                    <div className="flex items-center gap-1">
-                      <Highlighter
-                        className={`h-4 w-4 ${useHighlightedText ? "text-primary" : "text-muted-foreground"}`}
-                      />
-                      <span className={useHighlightedText ? "text-primary font-medium" : ""}>Highlighted Text</span>
-                    </div>
-                  </label>
-                </div>
-                <span>Press Enter to send, Shift+Enter for new line</span>
-              </div>
-            </div>
           </div>
         </CardContent>
       </Card>
+
+      {/* Citations Section */}
+      <Card className="border shadow-sm flex-1 flex flex-col overflow-hidden">
+        <CardHeader className="p-4 pb-2">
+          <CardTitle className="text-lg font-medium">Citations</CardTitle>
+        </CardHeader>
+        <CardContent className="p-4 pt-0 flex-1 overflow-hidden">
+  <ScrollArea className="h-full pr-3">
+    {usedCitations.length > 0 ? (
+      <div className="space-y-3">
+        {usedCitations.map((citation) => {
+          let modifiedDescription = citation.description;
+          const pageMatch = modifiedDescription.match(/, Page (\d+(\.\d+)?)/i);
+
+          if (pageMatch) {
+            const pageNumber = pageMatch[1];
+            modifiedDescription = modifiedDescription.replace(pageMatch[0], `#page=${parseInt(pageNumber) + 1}`);
+          }
+
+          return (
+          <div
+            key={citation.id}
+            className="border rounded-lg p-3 hover:bg-muted/50 transition-colors"
+          >
+            <div className="flex items-center justify-between mb-2">
+              <div className="bg-primary/10 text-primary text-xs font-medium px-2 py-1 rounded">
+                Citation #{citation.id}
+              </div>
+              {citation.description.startsWith("http") && (
+                          <a
+                            href={modifiedDescription}
+                            target="_blank"
+                            rel="noopener noreferrer"
+                            className="text-primary hover:text-primary/80"
+                          >
+                            <ExternalLink className="h-4 w-4" />
+                          </a>
+                        )}
+            </div>
+
+            {/* Truncate Long Links */}
+            <div className="text-sm w-full truncate">
+              
+              {citation.description.startsWith("http") && !citation.isimg && !citation.image_data ? (
+                <a
+                  href={modifiedDescription}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="text-primary hover:underline w-full block truncate"
+                  style={{ display: "block", maxWidth: "100%" }}
+                  title={citation.description} // Show full link on hover
+                >
+                  {modifiedDescription.length > 50
+                    ? modifiedDescription.slice(0, 50) + "..."
+                    : modifiedDescription}
+                </a>
+              ) : (<></>)}
+            </div>
+
+            {/* Fixed Image Size */}
+            {citation.isimg && citation.image_data && (
+              <div className="mt-2 bg-white p-2 rounded border flex justify-center">
+                <img
+                  src={`data:image/png;base64,${citation.image_data}`}
+                  alt={`Citation ${citation.id}`}
+                  className="rounded-md object-contain"
+                  style={{ maxWidth: "100%", maxHeight: "150px" }} // Adjust height to fit properly
+                />
+              </div>
+            )}
+          </div>
+        )})}
+      </div>
+    ) : (
+      <p className="text-muted-foreground text-center">No citations available</p>
+    )}
+  </ScrollArea>
+</CardContent>
+
+      </Card>
+
+      {/* Display the Graph in a Pop-up or Modal */}
+      {isGraphVisible && graphImage && (
+        <div className="fixed inset-0 bg-black bg-opacity-50 flex items-center justify-center z-50">
+          <div className="bg-white p-4 rounded-lg max-w-lg w-full relative">
+            {/* Close button positioned near the image */}
+            <Button
+              variant="ghost"
+              size="sm"
+              onClick={toggleGraphVisibility}
+              className="absolute top-2 right-2 text-muted-foreground"
+            >
+              <X className="h-5 w-5" />
+            </Button>
+            <div className="flex justify-center">
+              <img
+                src={`data:image/png;base64,${graphImage}`}
+                alt="Generated Graph"
+                className="rounded-md max-w-full"
+              />
+            </div>
+          </div>
+        </div>
+      )}
     </div>
   )
 }
